Use a constant port for the local-proxy
Opening a Server in the parent process attaches a listener to the server
'accept', this results in a possibility of the child process never
receiving some of the requests.

Closing the server in the parent process keeps the server open in the
child process, but prevents from passing the already closed server to
new child processes.

This limits the functionality to use only one port for development
(multiple npm run start will fail), but fixes #36 unpredictable
behavior.