Generate client, server from interfaces
This is an interim solution!

Rush and Concord do not play nicely together.  Create *fake* packages
for client and server definitions: `interfaces-{client,server}` link
during `rush update` as though they were the real thing, but after
build they forward to the generated code.

Problems include that dependencies from the generated code are not
propagated; we shall have to copy the really-required ones out from
`interfaces-{client,server}/gen/package.json` up to
`interfaces-{client,server}/package.json`.

This is an interim solution!

The right way to do it may be to change Concord to support building a
package from inside it.